Rename api files and directories, add manifest.dev.yml to gitignore
- /utils => /services
- IBMCloudIam.js => appId.js
