+++ web/.gitignore
@@ -1,2 +1,2 @@
 /node_modules
-manifest.yml
+manifest.dev.yml
